etcd: add PersistentLease and WatchPrefix (#123)
PersistentLease is a foundational component for service registries
and cluster memeberships. This will be useful in the future for
coordination, such as messaging and/or distributed NoncePools.

WatchPrefix is the read aspect of this, where we watch a set of
KVs under a prefix (in a snapshotting fashion) to be able to read
members of a cluster.